requirements.txt
flask
python-dotenv
openai
flask-cors
